chore: Index Reader clean up
